refactor: move exceptions to their namespace
